refactor(as7262): split Led functionality in new file
